+++ ui/types/formSchemas.ts
@@ -142,7 +142,13 @@ export const addCredentialsFormSchema = (providerType: string) =>
                     .nonempty("Client Secret is required"),
                   tenant_id: z.string().nonempty("Tenant ID is required"),
                   user: z.string().nonempty("User is required"),
+<<<<<<< HEAD
                   password: z.string().nonempty("Password is required"),
+=======
+                  password: z
+                    .string()
+                    .nonempty("Encrypted Password is required"),
+>>>>>>> 36aaec8a5 (chore(m365powershell): manage encryption from plaintext password (#7784))
                 }
               : {}),
   });